Move PureElement` to `pureDom`
